Initial commit. We are going to have a three-level architecture:
    noderep maps directly to storage
    zorba_node contains nodereps, and inherits from
    abstract_node

The point is to have zorba_nodes that use rtti, and nodereps that use placement new.   There is no way, in practice, to mix these two models.

This means that we have iterators that allocate and deallocate zorba_nodes containing noderep pointers in response to each call to next().  The system then needs to run a completely independent zorba_node cache to avoid excess allocation and deallocation of heap-allocated zorba_nodes.

Additional implementations can then implement abstract_node with references to other storage systems.  Again, these implementations are responsible for their own allocation optimization.

This  architecture is unfortunate, but it is the only way to reconcile the need for complete abstraction with the objective of having at least one native storage implementation.