Removed the following static libraries: zorbatypes, zorbautil, zorbaerrors. We now generate one zorba library for each store passed using the ZORBA_STORE_DIRS variable. This variable may contain a list of directories that contain different store implementations. The CMakeLists.txt must specify a ZORBA_STORE_IMPL_SRCS variable that lists all the sources of that store. The name of the library will be zorba_${ZORBA_STORE_LIB_EXTENSION} where ZORBA_STORE_LIB_EXTENSION can be defined in the CMakeLists.txt file that belongs to the store being compiled.
Also this change sets the default visiblity of library symbols to hidden. For Windows this is a necessity anyway. For Unix (GNU compiler) this gives us a little bit of performance when loading the library. Only symbols/classes marked with the preprocessor variable ZORBA_DLL_PUBLIC will be exported in the library. At the moment these are all classes defined in the include/zorba and include/simplestore directories plus some additional symbols that are used for unit testing (actually the ones used for unit testing are quite a lot and we might think of removing some here).

In order to build the minimal store, there is a shortcut: simply call cmake -D ZORBA_MINIMAL_STORE=ON. This will build the store located in src/store/minimal. If the ZORBA_STORE_DIRS variable is not set, the default simplestore located in src/store/naive will be used (resulting in a library called zorba_simplestore). Building the simplestore also triggers our rbkt and unit tests as well as the zorba command line utility to be built.