Added 3 more functions to the file module: - path-to-uri($path) - mkdir($dir, $create) - mkdirs($dir, $create)
path-to-uri transforms a path to a file:/// uri. The implementation is flexible and tolerates '\' and '/' on WIN32. If a URI is provided, this is validated and returned if valid, otherwise an error is thrown.
For mkdir functions, the $create boolean flag controls if the functions return false if a dir already exist at the given path. This simplifies the XQuery code in some cases (not needing multiple if expressions and exists test)