added the tidy functions into a new module (http://www.zorba-xquery.com/modules/tidy) that appears in the xqDoc documentation (the module is covered by the existing/modified) tidy tests.
The tidy function in the old namespace (http://www.zorba-xquery.com/zorba/util-functions)
are still available but not documented. Also, the tdoc functions in this namespaces
are not yet available in the new namespace. The reason therefore its unclear whether
this function should be declared as sequential because it stores the document in the store
(i.e. it has side effects)