Added test mainModule21 for bug # 3033534. The test is currently marked as expected failure.
The spec says the following:

It is a static error [err:XQST0036] to import a module if the importing module's in-scope schema types do not include definitions for the schema type names that appear in the declarations of variables and functions (whether in an argument type or return type) that are present in the imported module and are referenced in the importing module.

Zorba currently ignores the last words, i.e. it always requires the schema to be imported even if the function or variable is not referenced in the importing module. The test zorba/modules/mainModule21 has been added to the repository and is currently marked as expected failure.