Integrated Markus' socket stream buffer into the debugger server.
NOTE: The streambuf is synchronized on Windows and can generate deadlocks in the protocol if the same streambuffer is used for both input and output operations. The debugger server currently builds 2 stream buffers on the same socket, one used with an istream and the other with a ostream.

Added the FindSwig.cmake Windows proxy cmake module file (forgot to add it in my last commit)