Aligned names of the breakpoint commands to be similar to the ones specified in the protocol. bget for breakpoint_get, etc
all breakables are saved as URI's. Some commands (blist, i.e. breakpoint_list) will translate them back to paths before returning them to the client. (This only happens for the commands where the protocol does not enforce a URI in the returned filename attribute)

Refactored building of stack frames in the server.